+++ Modules/Validator/date.js
@@ -1,9 +1,11 @@
-import { standardValidation, isDate, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import isDate from 'front-library/Modules/Validator/Tools/isDate';
+import { addValidator } from 'front-library/Modules/Validator';
 
 /**
  * Date validation
  */
-addValidator( 'date', '[data-date-control]', ( $input, value ) => {
+addValidator( 'date', '[data-date-control]', ( $input, value, isLiveValidation ) => {
     return standardValidation(
         $input,
         value,
@@ -12,6 +14,8 @@ addValidator( 'date', '[data-date-control]', ( $input, value ) => {
                 value,
                 $input.getAttribute( 'data-date-control' )
             ),
-        'date'
+        'date',
+        undefined,
+        isLiveValidation
     );
 } );