+++ Modules/Validator/email.js
@@ -1,13 +1,17 @@
-import { standardValidation, isEmail, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import isEmail from 'front-library/Modules/Validator/Tools/isEmail';
+import { addValidator } from 'front-library/Modules/Validator';
 
 /**
  * Email validation
  */
-addValidator( 'email', '[type="email"]', ( $input, value ) => {
+addValidator( 'email', '[type="email"]', ( $input, value, isLiveValidation ) => {
     return standardValidation(
         $input,
         value,
         value === '' || isEmail( value ),
-        'email'
+        'email',
+        undefined,
+        isLiveValidation
     );
 } );