+++ Modules/Validator/min.js
@@ -1,14 +1,17 @@
-import { standardValidation, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import { addValidator } from 'front-library/Modules/Validator';
 
 /**
  * Min range validation
  */
-addValidator( 'min', '[min]', ( $input, value ) => {
+addValidator( 'min', '[min]', ( $input, value, isLiveValidation ) => {
     return standardValidation(
         $input,
         value,
         value === '' ||
             parseFloat( value ) >= parseFloat( $input.getAttribute( 'min' ) ),
-        'min'
+        'min',
+        undefined,
+        isLiveValidation
     );
 } );