+++ Modules/Validator/minlength.js
@@ -1,9 +1,10 @@
-import { standardValidation, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import { addValidator } from 'front-library/Modules/Validator';
 
 /**
  * Min length validation
  */
-addValidator( 'minlength', '[minlength]', function( $input, value ) {
+addValidator( 'minlength', '[minlength]', function( $input, value, isLiveValidation ) {
     var min;
 
     min = $input.getAttribute( 'minlength' );
@@ -12,6 +13,8 @@ addValidator( 'minlength', '[minlength]', function( $input, value ) {
         $input,
         value,
         value.length >= min,
-        'min'
+        'min',
+        undefined,
+        isLiveValidation
     );
 } );