+++ Modules/Validator/pattern.js
@@ -1,13 +1,16 @@
-import { standardValidation, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import { addValidator } from 'front-library/Modules/Validator';
 
 /**
  * Email validation
  */
-addValidator( 'pattern', '[pattern]', ( $input, value ) => {
+addValidator( 'pattern', '[pattern]', ( $input, value, isLiveValidation ) => {
     return standardValidation(
         $input,
         value,
         value === '' || new RegExp( $input.getAttribute( 'pattern' ) ).test( value ),
-        'pattern'
+        'pattern',
+        undefined,
+        isLiveValidation
     );
 } );