+++ Modules/Validator/recaptcha.js
@@ -1,10 +1,13 @@
-import { standardValidation, addValidator } from 'front-library/Modules/Validator';
+import { standardValidation } from 'front-library/Modules/Validator/Tools/ValidationState';
+import { addValidator } from 'front-library/Modules/Validator';
 
-addValidator( 'recaptcha', '.js--recaptcha', true, ( $input, value ) => {
+addValidator( 'recaptcha', '.js--recaptcha', true, ( $input, value, isLiveValidation ) => {
     return standardValidation(
         $input,
         value,
         !window.grecaptcha || window.grecaptcha.getResponse() !== '',
-        'recaptcha'
+        'recaptcha',
+        undefined,
+        isLiveValidation
     );
 } );